% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{amp_eff_output_extract}
\alias{amp_eff_output_extract}
\title{Extract amplification efficiencies from Stan Model 3}
\usage{
amp_eff_output_extract(stan_data, amp_eff_param)
}
\arguments{
\item{stan_data}{A `list` of data formatted as required by Stan Model 3, typically 
produced by the `prep_stan_M3` function.}

\item{amp_eff_param}{A `list` or `data.frame` containing the amplification efficiency 
parameters estimated by Stan Model 3. Typically, this includes the mean and standard 
deviation (`sd`) of the amplification efficiency estimates.}
}
\value{
A `data.frame` containing the following columns:
  - `Pre-PCR`: The pre-PCR relative abundances of each species in the mock community.
  - `Post-PCR`: The post-PCR mean sequencing counts for each species.
  - `Post-PCR_est`: The estimated post-PCR values based on the calculated amplification efficiency.
  - `amp_eff`: The amplification efficiency for each species, calculated as `exp(40 * mean(amp_eff_param))`.
  - `amp_eff_lo`: The lower bound of the amplification efficiency, based on the standard deviation.
  - `amp_eff_up`: The upper bound of the amplification efficiency, based on the standard deviation.
}
\description{
This function extracts the amplification efficiencies calculated by Stan Model 3 
and formats the results for further analysis. The output includes the pre-PCR and 
post-PCR values, as well as the estimated amplification efficiencies with their 
associated confidence intervals.
}
\examples{
# Extract amplification efficiencies from the Stan Model 3 output
amp_eff_output <- amp_eff_output_extract(stan_data = stan_data_M3, amp_eff_param = amp_eff_param)
}
