% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{amp_eff_param_extract}
\alias{amp_eff_param_extract}
\title{Extract Amplification Efficiency Parameters}
\usage{
amp_eff_param_extract(stanMod)
}
\arguments{
\item{stanMod}{A `stanfit` object containing the results of the model fitting.}
}
\value{
A data frame containing the summary statistics of the amplification efficiency parameters.
}
\description{
This function extracts amplification efficiency parameters from a fitted Stan model.
}
\examples{
# Extract amplification efficiency parameters
amp_eff_param <- amp_eff_param_extract(stanMod)
}
