% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{est_ini_prop_extract}
\alias{est_ini_prop_extract}
\title{Extract Initial Proportions}
\usage{
est_ini_prop_extract(stanMod)
}
\arguments{
\item{stanMod}{A `stanfit` object containing the results of the model fitting.}
}
\value{
A data frame representing the initial proportional abundances of species.
}
\description{
This function extracts the initial proportional abundances for species in a mock community
from a fitted Stan model.
}
\examples{
# Extract initial proportions
est_ini_prop <- est_ini_prop_extract(stanMod)
}
