% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{est_ss_quant_extract}
\alias{est_ss_quant_extract}
\title{Extract estimated eDNA qPCR concentrations from model M2}
\usage{
est_ss_quant_extract(stanMod)
}
\arguments{
\item{stanMod}{A `stanfit` object containing the results of fitting Stan Model 2, typically produced by `run_M2`.}
}
\value{
A `data.frame` (tibble) with the following columns:
  - `sample_index`: The index of the sample as used in the Stan model.
  - `Sample_name`: The name or identifier of the sample which was inputed in Stan model.
  - `C_est_log`: The mean of the log10-transformed estimated DNA concentrations.
  - `C_est_log_se`: The standard error of the log10-transformed estimated DNA concentrations.
  - `C_est_log_sd`: The standard deviation of the log10-transformed estimated DNA concentrations.
}
\description{
This function extracts the estimated quantities (e.g., DNA concentrations) from the fitted Stan Model 2.
It processes the model output to return a data frame containing the sample indices, sample names, and
the estimated log10-transformed DNA concentrations along with their associated standard errors and standard deviations.
}
\examples{
# Extract the estimated quantities from the fitted Stan Model 2
est_ss_quant <- est_ss_quant_extract(stanMod = stanMod_2)
}
