% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{extract_ini_prop}
\alias{extract_ini_prop}
\title{Extract Initial Proportions}
\usage{
extract_ini_prop(model_output)
}
\arguments{
\item{stanMod}{A `stanfit` object containing the results of the model fitting.}
}
\value{
A data frame representing the initial proportional abundances of species.
}
\description{
This function extracts the initial proportional abundances for species in a mock community
from a fitted Stan model.
}
\examples{
# Extract initial proportions
est_ini_prop <- extract_ini_prop(stanMod)
}
