% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{extract_matrix}
\alias{extract_matrix}
\title{Extract 2-Dimensional Model parameters as a matrix}
\usage{
extract_matrix(stanMod, param, vector = "mean")
}
\arguments{
\item{stanMod}{A `stanfit` object containing the results of the model fitting.}

\item{param}{A `character` string specifying the name of the 2-dimensional parameter to be extracted
from the model (e.g., `"alpha"` for `alpha[m,n]`).}

\item{vector}{A `character` string specifying the summary statistic to be extracted for each element
of the matrix. Options might include `"mean"`, `"se_mean"`, `"sd"`, or quantiles (e.g., `"2.5%"`).
Defaults to `"mean"`.}
}
\value{
A `data.frame` representing the 2-dimensional parameter as a matrix, with rows and columns
  corresponding to the indices of the parameter.
}
\description{
This function extracts and organizes 2-dimensional parameters (e.g., `alpha[m,n]`)
from a fitted Stan model into a matrix format. The user can specify which summary statistic
(e.g., mean, standard error) to extract.
}
\examples{
# Extract the mean values of the 2-dimensional parameter "alpha[m,n]" as a matrix
alpha_matrix <- extract_matrix(stanMod = stanMod_1, param = "alpha", vector = "mean")

# Extract the standard deviations of the 2-dimensional parameter "beta[m,n]" as a matrix
beta_matrix <- extract_matrix(stanMod = stanMod_1, param = "beta", vector = "sd")
}
