% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{extract_qpcr_param}
\alias{extract_qpcr_param}
\title{Extract Model Parameters from Stan Output (M1 & M2)}
\usage{
extract_qpcr_param(model_output)
}
\arguments{
\item{stanMod}{A `stanfit` object containing the results of the model fitting,
typically produced by `run_M1`.}
}
\value{
A `data.frame` containing the extracted parameters: `logit_phi`, `beta_0`, `beta_1`, `gamma_0`, and `gamma_1`.
}
\description{
This function extracts specific model parameters from the Stan model M1 and M2 output.
}
\examples{
# Assuming you have run the model and have a stanfit object called stanMod_1
ss_param <- extract_qpcr_param(stanMod = stanMod_1)
}
