% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{load_model}
\alias{load_model}
\title{Load and Compile a Stan Model}
\usage{
load_model(model = "M1")
}
\arguments{
\item{model}{A character string specifying which model to load. The available options are
`"M1"`, `"M2"`, `"M3"`, `"M4"`, `"M5_1"`, and `"M5_2"`. Default is `"M1"`.}
}
\value{
An object of class `stanmodel`, representing the compiled Stan model. This object can be used
directly in `rstan` functions such as `sampling`.
}
\description{
The `load_model` function loads and compiles a specified Stan model from a predefined list of models.
The models are written as character strings in the R environment, and this function compiles them into
a Stan model object using the `rstan::stan_model` function.
}
\details{
The function looks up the provided model name in a predefined list that includes the Stan code
and model name. If the model name is not found, the function throws an error. If the model name is valid,
the corresponding Stan model code is compiled and returned as a `stanmodel` object.
}
\examples{
\dontrun{
# Load and compile the M1 model
M1 <- load_model("M1")

# Load and compile the M3 model
M3 <- load_model("M3")

# Load and compile the M5.2 model
M5_2 <- load_model("M5_2")
}

}
