% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{mock_variance_extract}
\alias{mock_variance_extract}
\title{Extract Mock Variance}
\usage{
mock_variance_extract(stan_data, stanMod, amp_eff_param)
}
\arguments{
\item{stan_data}{A list containing the mock community data formatted for Stan.}

\item{stanMod}{A `stanfit` object containing the results of the model fitting.}

\item{amp_eff_param}{A data frame containing amplification efficiency parameters.}
}
\value{
A data frame containing the differences in proportions between modeled and observed values.
}
\description{
This function calculates the difference between observed and modeled proportional abundances
for mock community samples.
}
\examples{
# Extract mock variance
mock_variance <- mock_variance_extract(stan_data, stanMod, amp_eff_param)
}
