% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{mutate_indices}
\alias{mutate_indices}
\title{Group and Add Index to a Data Frame}
\usage{
mutate_indices(data, index_variable, index_name)
}
\arguments{
\item{data}{A data frame to be processed.}

\item{index_column}{A character string specifying the name of the column to group by.}
}
\value{
A data frame with an added column `index` that assigns a unique ID to each group.
}
\description{
This function groups a data frame by a specified column and adds an index to each group.
}
\examples{
# Example usage
df <- data.frame(group = c("A", "A", "B", "B", "C"), value = 1:5)
df \%>\% mutate_indices('group', 'i_idx')
}
