% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{plot_est_ss_quant}
\alias{plot_est_ss_quant}
\title{Plot estimated eDNA qPCR concentrations from model M2}
\usage{
plot_est_ss_quant(est_ss_quant, scale = "log")
}
\arguments{
\item{est_ss_quant}{A `data.frame` containing the estimated DNA concentrations and their standard deviations derived from `est_ss_quant_extract` function.}

\item{scale}{A `character` string specifying whether the y-axis should be on a logarithmic scale. 
Defaults to `"log"`. If `scale = "log"`, the y-axis will be displayed on a logarithmic scale; 
otherwise, a linear scale will be used.}
}
\value{
A `ggplot` object showing the estimated DNA concentrations with error bars.
  If `scale = "log"`, the y-axis is scaled logarithmically and labeled using scientific notation.
}
\description{
This function creates a plot of estimated DNA concentrations for single-species qPCR,
including error bars representing the uncertainty (standard deviation) around each estimate.
The y-axis can be displayed on a logarithmic scale if specified.
}
\examples{
# Plot the estimated DNA concentrations with a logarithmic y-axis
plot_est_ss_quant(est_ss_quant = est_ss_quant, scale = "log")

# Plot the estimated DNA concentrations with a linear y-axis
plot_est_ss_quant(est_ss_quant = est_ss_quant)
}
