% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{plot_mock_variance}
\alias{plot_mock_variance}
\title{Plot Mock Variance}
\usage{
plot_mock_variance(mock_variance)
}
\arguments{
\item{mock_variance}{A data frame containing differences between modeled and observed proportions.}
}
\value{
A `ggplot` object showing the variance for each species across samples.
}
\description{
This function generates a plot to visualize the variance between modeled and observed
proportional abundances for mock community samples.
}
\examples{
# Plot mock variance
plot_mock_variance(mock_variance)
}
