% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{plot_sample_variance}
\alias{plot_sample_variance}
\title{Plot Sample Variance}
\usage{
plot_sample_variance(sample_variance)
}
\arguments{
\item{sample_variance}{A data frame containing differences between modeled and observed proportions.}
}
\value{
A `ggplot` object showing the variance for each species across samples.
}
\description{
This function generates a plot to visualize the variance between modeled and observed
proportional abundances for environmental samples.
}
\examples{
# Plot sample variance
plot_sample_variance(sample_variance)
}
