% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{plot_ss_param_pd}
\alias{plot_ss_param_pd}
\title{Plot Probability of Detection}
\usage{
plot_ss_param_pd(stan_data, ss_param, xmin_log, xmax_log)
}
\arguments{
\item{stan_data}{A `list` of data formatted for input into the Stan model,
produced by `prep_stan_M1`.}

\item{ss_param}{A vector of output (unknown) parameters produced by fitting 
the Stan model through `run_M1`. It can also be retrieved by `ss_param_extract`.}

\item{xmin_log}{A `numeric` value specifying the lower bound of the DNA concentration 
range to be plotted on a logarithmic scale. If not provided, it defaults to 
the floor of the minimum DNA concentration in `stan_data`.}

\item{xmax_log}{A `numeric` value specifying the upper bound of the DNA concentration 
range to be plotted on a logarithmic scale. If not provided, it defaults to 
the rounded maximum DNA concentration in `stan_data`.}
}
\value{
A `ggplot` object showing the probability of detection along the DNA concentration.
}
\description{
This function creates a plot of the probability of detection along the DNA concentration range,
based on the Stan model's output.
}
\examples{
plot_ss_param_pd(stan_data = stan_data_M1, ss_param = ss_param)
}
