% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{prep_stan_M2}
\alias{prep_stan_M2}
\title{Prepare data for Stan Model 2}
\usage{
prep_stan_M2(
  qpcr_data,
  sample_type,
  Ct,
  sample_name_column,
  standard_concentration,
  plate_index
)
}
\arguments{
\item{sample_type}{A `character` string specifying the name of the column in `data`
that indicates the type of sample (e.g., "STANDARD" for standard samples
and "UNKNOWN" for environmental samples).}

\item{Ct}{A `character` string specifying the name of the column in `data`
that contains the Ct values. qPCR non-detects should be defined as 'Undetermined'
and not NA or any other value.}

\item{sample_name_column}{A `character` string specifying the name of the column
in `data` that contains the unique sample names or identifiers (index).}

\item{standard_concentration}{A `character` string specifying the name of the
column in `data` that contains the standard concentrations for the standard samples.
This column should be numeric and have NA values for environmental samples. If not it will be converted to numeric within the function.}

\item{data}{A `data.frame` containing the qPCR data. It should include the
Ct values, sample types (e.g., "STANDARD", "UNKNOWN"), sample names, and
standard (initial) concentrations.}
}
\value{
A list formatted as required by the Stan model.
}
\description{
This function prepares the data required to run Stan Model 2 by filtering
the input data, selecting the necessary columns, and formatting the data
to include both standard and unknown (environmental) qPCR samples.
}
