% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{run_M1}
\alias{run_M1}
\title{Run Stan Model 1}
\usage{
run_M1(stan_object = M1, stan_data = stan_data_M1, plot_fig = T)
}
\arguments{
\item{stan_object}{A `stanmodel` object representing the compiled Stan model (M1).}

\item{stan_data}{A `list` of data formatted for input into the Stan model,
produced by `prep_stan_M1`.}

\item{plot_fig}{A `logical` value indicating whether to generate plots of the
model's outputs after fitting the model. Defaults to `TRUE`. If `TRUE`, the 
function will generate and display the output plots; if `FALSE`, no plots 
will be produced.}
}
\value{
An object of class `stanfit` containing the results of the model fitting.
The model runs 4 chains, 5000 iterations and burns 2000 for warm-up. For changing these values run the analog version of this function
}
\description{
This function runs Stan Model 1 using the prepared data and optionally 
generates plots of the model's outputs.
}
\examples{
stanMod_1 <- run_M1(stan_object = M1, stan_data = stan_data_M1)
}
