% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{Run_Model}
\alias{Run_Model}
\title{Run Stan Model}
\usage{
Run_Model(
  stan_object,
  stan_data,
  iterations = 5000,
  warmup = 2000,
  no_chains = 4,
  treedepth = 10
)
}
\arguments{
\item{stan_object}{A `stanmodel` object representing the compiled Stan model (M2).}

\item{stan_data}{A `list` of data formatted for input into the Stan model,
produced by `prep_stan` function.}
}
\value{
An object of class `stanfit` containing the results of the model fitting.
The model runs 4 chains, 5000 iterations and burns 2000 for warm-up. For changing these values run the analog version of this function
}
\description{
This function runs Stan Model using the prepared data to estimate the environmental DNA (eDNA) concentrations
of unknown environmental samples that have been processed through qPCR alongside standard samples of known concentrations.
}
\examples{
stanMod_2 <- run_M2(stan_object = M2, stan_data = stan_data_M2)
}
