% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{sample_variance_extract}
\alias{sample_variance_extract}
\title{Extract Sample Variance}
\usage{
sample_variance_extract(est_ini_prop, amp_eff_param, stan_data, stanMod)
}
\arguments{
\item{est_ini_prop}{A data frame of initial proportional abundances extracted from the Stan model.}

\item{amp_eff_param}{A data frame of amplification efficiency parameters.}

\item{stan_data}{A list containing the environmental sample data formatted for Stan.}

\item{stanMod}{A `stanfit` object containing the results of the model fitting.}
}
\value{
A data frame containing the differences in proportions between modeled and observed values for samples.
}
\description{
This function calculates the difference between observed and modeled proportional abundances
for environmental samples.
}
\examples{
# Extract sample variance
sample_variance <- sample_variance_extract(est_ini_prop, amp_eff_param, stan_data, stanMod)
}
