% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{ss_param_extract}
\alias{ss_param_extract}
\title{Extract Model Parameters from Stan Output (M1 & M2)}
\usage{
ss_param_extract(stanMod)
}
\arguments{
\item{stanMod}{A `stanfit` object containing the results of the model fitting,
typically produced by `run_M1`.}
}
\value{
A `data.frame` containing the extracted parameters: `alpha_0`, `alpha_1`, `eta_0`, `eta_1`, `gamma_0`, and `gamma_1`.
}
\description{
This function extracts specific model parameters from the Stan model M1 and M2 output.
}
\examples{
# Assuming you have run the model and have a stanfit object called stanMod_1
ss_param <- ss_param_extract(stanMod = stanMod_1)
}
