% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_functions.R
\name{prep_stan_M1}
\alias{prep_stan_M1}
\title{Prepare data for Stan Model 1}
\usage{
prep_stan_M1(data, Ct, standard_concentration)
}
\arguments{
\item{data}{A `data.frame` containing the qPCR data. It should include the 
Ct values and standard (initial) concentrations.}

\item{Ct}{A `character` string specifying the name of the column in `data` 
that contains the Ct values. qPCR non-detects should be defined as 'Undetermined' and not NA or any other value}

\item{standard_concentration}{A `character` string specifying the name of the 
column in `data` that contains the standard concentrations.}
}
\value{
A list formatted as required by the Stan model.
}
\description{
This function prepares the data required to run Stan Model 1 by filtering
the input data and selecting the necessary columns.
}
\examples{
stan_data_M1 <- prep_stan_M1(data = qpcr \%>\% filter(task == "STANDARD"),
                             Ct = "Ct",
                             standard_concentration = "st_conc")
}
